Feat: Add support for two additional LEDs (led:torch_2 and led:torch_3) (#39)
* Fix selinux

* Control secondary flash LEDs in light and brightness cycles

This change modifies the light and brightness cycle functionality in the ExperimentalActivity to control both the primary and secondary flash LEDs.

The `controlLeds` function is updated to include brightness values for the secondary LEDs (WHITE2_LED_PATH, YELLOW2_LED_PATH), ensuring both primary and secondary LEDs are controlled during the cycles.

* Bump version to 1.3.2

- Update `versionCode` from 28 to 29.
- Update `versionName` from 1.3.1 to 1.3.2.